Authentication for Livewire components

+++ app/Http/Livewire/PackageRoom.php
@@ -4,9 +4,12 @@ namespace App\Http\Livewire;
 
 use App\Models\Package;
 use Livewire\Component;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class PackageRoom extends Component
 {
+    use AuthorizesRequests;
+
     public $room;
 
     public function render()
@@ -22,14 +25,19 @@ class PackageRoom extends Component
         return $this->room->venue->packages()->whereNotIn('id', $this->room->packages->pluck('id'))->get();
     }
 
+    // TODO: Is this the correct permission?
     public function add(Package $package)
     {
+        $this->authorize('modify rooms');
+
         $this->room->packages()->attach($package->id);
         $this->room->refresh();
     }
 
     public function remove(Package $package)
     {
+        $this->authorize('modify rooms');
+
         $this->room->packages()->detach($package->id);
         $this->room->refresh();
     }
